Fixed textasset loading on android, swapped cameras for XR rigs
Needed to make the TextManager a proper singleton, and implement a new 'initialization' scene where it can load all the text databases before the rest of the game loads. This is because android requires unitywebrequest coroutine instead of just using file.load, etc. What we could do if we don't want this initialization scene in the future is have proper references to these text assets, maybe make them scriptable objects event.