Fixing the repeated sound bug once and for all. Added Unity AudioMixer that all sounds now use.
Accomplished this by muting the whole game while we throttle unused asssets. Also added a setting in [GUI] section of settings.ini to disable garbage collection when closing UI windows, called 'DisableGarbageCollectionDuringMenuing'. Users can flag this 'true' if the occasional muting of the game is too much to bear, but it's recommended to keep 'false' since otherwise the game will be more prone to crashing.