Implemented keycode mapping for touchscreen buttons
This complements the preexisting action mapping. A button can have one or the other or even both active simultaneously. Keycodes are useful for mods that haven't implemented 'CustomActions' (and now that I've implemented this and fixed the keyboard input for mod settings, I might do away with Custom actions altogether)