Added 'AndroidSimulationManager' to make it easier to develop using the android simulator. Fixed render texture not rendering UI masks correctly.
Currently, all the new manager does is change the resolution of everything to match the current game view, regardless of if we're in Simulator mode or Game mode. As for the render texture and UI masks, this was affecting the dropdowns, causing them to render outside of the area we wanted them to render.